feat: centralized connection added

+++ contracts/sui/xcall/sources/connections.move
@@ -24,7 +24,7 @@ const ConnCentralized:vector<u8> =b"centralized";
         
     }
 
-    public fun get_fee(states:&mut Bag,package_id:String,netId:String,response:bool):u128{
+    public fun get_fee(states:&mut Bag,package_id:String,netId:String,response:bool):u64{
 
          if (package_id==centralized_connection::package_id_str()){
             let fee= centralized_connection::get_fee(states,netId,response);
@@ -37,6 +37,5 @@ const ConnCentralized:vector<u8> =b"centralized";
        
 
     }
-
     
 }
